pyproject.toml
[tool.poetry]
name = "daikiaxis"
version = "0.1.0"
description = ""
authors = ["Your Name <c5sc3.gri.n4w7@gamil.com>"]

[tool.poetry.dependencies]
python = "3.8.9"
jupyterlab = "^3.1.12"
numpy = "^1.21.2"
pandas = "^1.3.3"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
pandas-profiling = "^3.0.0"
ipywidgets = "^7.6.5"
jupyterlab-widgets = "^1.0.2"
lightgbm = "^3.2.1"
msoffcrypto-tool = "^4.12.0"
scikit-learn = "^1.0.1"
graphviz = "^0.17"
openpyxl = "^3.0.9"
autopep8 = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
